Merge progress reporting sections
When I created the new section about progress reporting on the "Reporting" page, I'd overlooked that the "Usage" page already contained a section on progress reporting.

This commit now removes the section from the "Reporting" page again and merges some of the information which was contained in that section, into the information about progress reporting on the "Usage" page.

Includes various tweaks and updates to the pre-existing information:
* Make it explicit how progress reporting can be enabled.
* Shorten the example output for progress.
* Mention the name of the configuration option to enable progress reporting by default.
* Refresh the example output for the verbose example.